+++ CHANGELOG.md
@@ -51,6 +51,7 @@ All notable changes to this project will be documented in this file.
 - Remove headings without content from _References_ tab in Main View
 - HTML content in comments
 - Adding comments without content to _Entity_
+- Storing _false_ values in _Boolean_ attribute
 ### Changed
 - _Entity tree_ sorting is now accessible through ...-menu
 - Reset Password Workflow